+++ src/main/java/nextstep/auth/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.auth.support;
+package nextstep.auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
