+++ src/main/java/nextstep/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package nextstep.auth.support;
+package nextstep.auth.exception;
 
 public class InvalidTokenException extends RuntimeException{
 }
