+++ src/main/java/nextstep/reservation/service/ReservationService.java
@@ -7,8 +7,8 @@ import nextstep.reservation.dao.ReservationDao;
 import nextstep.reservation.model.ReservationRequest;
 import nextstep.schedule.model.Schedule;
 import nextstep.schedule.dao.ScheduleDao;
-import nextstep.auth.support.AuthenticationException;
-import nextstep.support.DuplicateEntityException;
+import nextstep.auth.exception.AuthenticationException;
+import nextstep.exception.DuplicateEntityException;
 import nextstep.theme.model.Theme;
 import nextstep.theme.dao.ThemeDao;
 import org.springframework.stereotype.Service;
